Improve Node parent validators.
Add a second argument for the node itself, so the validator for
circular references doesn't have to capture. This capture was leading
do a dangling this pointer when a Node was copy-constructed.

Change the validators from functions which return lambdas to regular
functions which do the validation and return bool. We can now add
these functions directly as validators rathe than calling them and
returning the result. If nothing else this makes gdb happier.