Another pass at the Table API.
Remove the wrapper class and make Table a typedef for a variant.

Add more types at the Table level, including vectors as well as maps,
and remove some variants from the inner types. In many cases only a
single boost::get is required to get to a homogeneous container, so
this is a bit easier to use.