Improve getData/setData interface in ByteMap and ImageByteMap.
Remove all convenience methods from ByteMap, leaving only the
byte-array interface. This class is not exposed directly to users so a
minimal interface is appropriate.

In ImageByteMap, replace the various getData/setData overloads with a
pair of templates which can handle any POD type. This relies on the
ability to access the byte representation of any POD object by
converting to gsl::span<gsl::byte>.