Use variable-sized allocations in ByteMap.
Rather than allocating fixed-sized pages, grow each region as data is
added, guaranteeing that contiguous data is allocated in a single
region.

This alters the semantics slightly: previously pages were zeroed on
allocation and access to those default-initialized regions was
allowed. Now access to data which has not been explicitly set will
throw.

Also, attempting to set data which partially overlaps an existing
region will throw. This restriction can be lifted in the future if it
proves onerous.