Initial commit changing the Node hierarchy.
This adds the notion of a Context object responsible for tracking
allocations, which will some day use a better allocator than malloc(),
of the various Node objects. It then removes the value semantics used
by the various Node subclasses (such as copy and move operations).

Patch is compiling cleanly with MSVC, but is untested due to the
testing system not being ready on that platform just yet.