Initial commit changing the Node hierarchy.
This adds the notion of a Context object responsible for tracking
allocations, which will some day use a better allocator than malloc(),
of the various Node objects. It then removes the value semantics used
by the various Node subclasses (such as copy and move operations).

Patch is compiling cleanly with MSVC, but is untested due to the
testing system not being ready on that platform just yet.

It also adds the cast/dyn_cast machinary from LLVM to GTIRB so that we can
remove the virtual functions from the Node hierarchy and instead use
cast/dyn_cast as safe casting operations. This is a WIP as the casting
machinery should not live in Context.hpp, but I wanted to put it
somewhere safe until I proved the Node hierarchy changes to be viable.

Finally, it removes the vtable from Node.

Note that some of the tests have been temporarily disabled while I
determine the best way to go about exposing the functionality again.