Make setUUID() into a virtual function.
This is needed so that we can clear UUIDs in a type-generic manner, but
I am not certain whether this is an implementation detail we expect to
keep exposed forever. When we disable RTTI and get rid of the dynamic
casts, this implementation can be replaced with CRTP if we want to keep
it exposed.