Add machinery for removing virtual inheritance.
This adds the cast/dyn_cast machinary from LLVM to GTIRB so that we can
remove the virtual functions from the Node hierarchy and instead use
cast/dyn_cast as safe casting operations. This is a WIP as the casting
machinery should not live in Context.hpp, but I wanted to put it
somewhere safe until I proved the Node hierarchy changes to be viable.