No longer initializing the Addr class's address.
There are no "safe" sentinel values to be used for an Addr object, but
removing the default constructor has too many downsides (like irregular
behavior with containers, use as data members in a class with a default
constructor, etc). Leaving the value uninitialized allows tools to warn
the user on uninitialized uses, so we're trying that.

Note that some of the tests are no longer valid because they're
assuming you can compare uninitialized values and get good behavior.