Update README.md to account for renaming EA -> Addr
Outstanding questions (with [line] locations):

[62-63] Should these elements map directly to class names, or are they
abstract names that are themselves mapped to classes? If the former,
we should change the entries here to match class names. If the latter,
we should provide the mapping.

[221] Should this be "`DataObject` or `InstructionRef`" ?

[240-241] I don't see a declaration for Block(Addr,Addr) in Block.hpp
or Block.cpp; this may need to be rewritten in terms of
Block(Addr,uint64_t,uint64_t)

[258] Clarify that add_edge() comes from Boost?

[263] Is this indexing allowed? don't you have to access the edge list first?

Also: should this be added to the Doxygen 'manual'? It seems more directly
relevant to the API than several of the other .md files in the manual.