Refactor the static UUID map to be in Context.
This moves the static UUID map into Context so that users can use a
mutex to protect the Context object itself when making multithreaded
use of gtirb. Any API that may modify this map requires a Context
object to be passed, and that object can be protected by a mutex
outside of gtirb if needed.