Helpfully add the protobuf source files to the IDE.
This allows IDEs like Visual Studio or Xcode to directly navigate to
the protobuf definition files. The added files are not part of the
compiled source for the project but are purely a navigational aid.

If we prefer they show up in the proto project rather than the gtirb
project, we can shift them to the other project, but this felt like the
most natural place for them to live.