Prevent double-linking of protobuf classes.
The generated code is linked into a static library, which CMake adds
to the interface of gtirb, causing the TestGTIRB executable to link
with it as well. This creates multiple definitions of the message
classes, which cause errors at runtime.

Fixed by making the proto lib a private dependency of gtirb.