Clean up the Node header file somewhat.
Remove the public functions that convert UUID to/from strings; they are
only used in Table.cpp and can be made private.

Move the definition of UUID into Context.hpp, where it already lived.
This also updates the documentation in Context.hpp accordingly.

Improved the documentation of the Node class.