More API doxygen work
Created several doxygen "groups" to gather together documentation for
related functionality that is not otherwise encapsulated.

Added \file entries to all files except those that are not originally
ours.

Added doxygen comment skeletons to namespaces gtirb, .

Added assorted \relates markings so that gtirb-level functions that
are related to specific classes are noted in the documentation for
those classes.