Add explicit constructor and assignment operators.
This is required to work around a bug in MSVC that causes a compile
error with the AuxDatas data member. The explicitly deleted copy
constructor and copy assignment operator are needed to silence a
spurious compile error, and the explicitly default move constructor and
move assignment operator are needed to allow the type to be moveable by
value.