Call node destructors when freeing the Context.
Imports SpecificBumpPtrAllocator from LLVM. This uses
BumpPtrAllocator (which we've already included), but only allocates
for a single type, and calls destructors for individual objects.

Context now has a separate allocator for each node type, with a
templated Allocate method which is specialized to use the correct
allocator.

This requires a bit of boilerplate when adding new node classes, but
as that's a rare event it shouldn't be too onerous.