Revert "Remove use of RTTI from AuxData."
This reverts commit 0f70aa9ece3608ab84a00bf3e2002b50af862acc.

This does not work reliably across shared library boundaries, although
various sources claim that it ought to. We end up with multiple
instantiations of the ID function which do not match.

This causes problems in ddisasm, where the generated IR is then passed
into the pretty-printer library for printing, at which point we can't
retrieve any AuxData.

As it turns out, we require RTTI anyway for protobuf, so there doesn't
seem to be much harm in using it for AuxData as well.