integrate protobuf to the build system.
If protobuf is not installed on the machine, then
pull/configure/install it in the local build directory so that gtirb
can use it.

Caveats: We cannot find_package(Protobuf..) after installing it a
custom location, so we need perform a limited number of things that
find_package(Protobuf..) does.