Change UUID codec, update Aux Data test
Changed the UUID codec to check the cache of Node UUIDs during decoding,
returning an object in the cache if it exists. This streamlines the use
of the standard AuxData tables. The updated AuxData test checks that
the instances of the data it contains are correct and not just UUIDs.