Update our cmake offerings a bit.
This adds the ability to control the number of parallel compile jobs
used when building with MSVC (it previously always did parallel
compilation without an option to control the number of processes
spawned). It also removes some configuration settings that were either
incorrect (like defaulting to MBCS instead of UNICODE) or unnecessary.
It also adds some extra flags when building in release mode with MSVC.