Update blocks/byte-intervals on changes to bytes
This is all done through the `bytes' method, which sets a range in a
vector of bytes designed by start and end indices to a supplied value.
This may be used to both delete and insert bytes at particular places
in a vector of bytes.  After the byte vector of a byte-block (i.e.,
code-block or data-block) or a byte-interval has changed the following
cleanup actions are performed:
- Update the size of blocks and byte-intervals whose bytes have
  changed
- Update the byte-interval when a block under it has changed
- Update the downstream blocks and symbolic-expressions of a
  byte-interval when its bytes have changed