Don't use __declspec(dllimport) on Windows
In "export" mode, we continue to use __declspec(dllexport),
but in "import" mode, just leave out the declspec altogether.
This allows gtirb to work both in with GTIRB_BUILD_SHARED_LIBS
either ON or OFF (i.e., dll mode or static mode).
Note that this works as long as we only export functions/methods
and not data.