Sanity check failure conditions when deserializing.
We previously would assume that we could load many parts of a serialized
GTIRB file without checking that the results made sense. For instance, we
did not check for null pointers, valid UUID formats, etc. Now we check for
these conditions and refuse to load the file instead of giving a user a known-
incorrect file they have to discover themselves.