Updates AuxData offsets on (setf bytes) in CL API
When the new *update-aux-data-offsets* variable is bound to true (off
by default because this behavior is slow) then on every setf of bytes,
go through and update all offsets in all AuxData tables that are
affected by the change in bytes.

This assumes that offsets are always using a data block as their
anchor -- which is the case in all observed existing AuxData tables.