WIP: Changes to AuxData schema for function names
This commit proposes some changes to our existing provisional AuxData
tables to support tools that try to identify libraries in binaries.
We need to be able to hold and refine multiple guesses from multiple
tools.  This commit merges the required new functionality with
existing tables for holding the names of required libraries.  Existing
tools like ddisasm and pprinter that use the existing "libraries"
table will need to be modified to use the new "libraryNames" table.