Add support for 8-byte signature to CL API
This introduces support for an 8-byte signature w/ magic and
version numbers to the GTIRB Common Lisp API.  As described
in the CHANGELOG.md, this change is not backwards compatible;
new versions of GTIRB will not be able to load older GTIRB file.
As a consequence, several tests relating to migrating older
GTIRB files to newer versions were removed as deprecated.