Update hello.gtirb to latest ddisasm output
This included the following unit test updates:

1. Update `address-range-block-lookup` to handle a list being
returned by `on-address`/`at-address`.  This necessitated adding
a `(mappend #'blocks)` wrapper after calling these functions.

2. Update key argument in `block-comments` to sort using the displacement
of an offset value (`[#'second #'car]`).

3. Update `payload-can-be-read-and-set` to remove testing of value symbols.
Unfortunately, no symbols of this type were present in the updated GTIRB,
and I felt generating a more complex GTIRB file which may or may not
contain such symbols risked breaking existing tests in unexpected ways.

Additionally, a minor modification was made to gtirb.lisp:

1. Modify `update-offset` to avoid attempting to find the `offset`
of an object which is not a `gtirb-byte-block`.  Previously, a crash
could occur when attempting to find the `offset` of a `byte-interval`
object.