Introduce JUnit and convert existing Java API tests to work with JUnit
I've removed use of the gtirb file for Java testing and instead have tests construct GTIRB IRs on the fly. All of the unit tests have been revamped (and actually perform tests rather than just producing stdout and "succeeding" by default.)

Also includes a change to the TreeMapUtils for range queries to make the upper bound of a range be exclusive rather than inclusive (which is more in line w/ the other APIs and more likely to be expected behavior for clients.) I've also updated comments range-based lookup functions to be explicit about inclusivity.